src/requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 科学计算和数据处理
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
Pillow>=9.0.0  # PIL库的Python实现

# 可视化相关
matplotlib>=3.5.0
seaborn>=0.11.0

# 其他工具
tqdm>=4.65.0  # 进度条显示
PyYAML>=6.0   # 配置文件处理

# 图像处理和增强
opencv-python>=4.7.0
albumentations>=1.3.0  # 图像增强库
